src/info21/requirements.txt
Django==4.2.10
python-dotenv==1.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
